removed unnecessary cast

+++ src/cmds/mpi_cmd_cpd.c
@@ -180,7 +180,7 @@ int splatt_mpi_cpd_cmd(
   cpd_cmd_args args;
   default_cpd_opts(&args);
   argp_parse(&cpd_argp, argc, argv, ARGP_IN_ORDER, 0, &args);
-  srand((int) args.opts[SPLATT_OPTION_RANDSEED]);
+  srand(args.opts[SPLATT_OPTION_RANDSEED]);
 
   sptensor_t * tt  = NULL;
   splatt_csf * csf = NULL;
